requirements.txt
Pillow==8.4.0
keyboard~=0.13.5
pystray==0.18.0